Move low integrity config class to a separate header.
