+++ SandboxTest/SandboxTest.vcxproj.filters
@@ -3,4 +3,7 @@
   <ItemGroup>
     <ClCompile Include="Main.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="LowIntegrity.hpp" />
+  </ItemGroup>
 </Project>
